separate component support exports
