+++ src/components/ui/toggle-group.tsx
@@ -5,7 +5,7 @@ import * as ToggleGroupPrimitive from "@radix-ui/react-toggle-group";
 import type { VariantProps } from "class-variance-authority";
 
 import { cn } from "~/lib/utils";
-import { toggleVariants } from "~/components/ui/toggle";
+import { toggleVariants } from "~/components/ui/component-variants";
 
 const ToggleGroupContext = React.createContext<
   VariantProps<typeof toggleVariants>
